Sprites are now in a vector instead of an index map, for simplicity

+++ roguelike_main/src/animation.rs
@@ -4,7 +4,7 @@ use roguelike_core::map::Aoe;
 use roguelike_core::config::Config;
 
 
-pub type SpriteKey = i64;
+pub type SpriteKey = usize;
 
 pub type SpriteIndex = f32;
